requirements.txt
python-telegram-bot==13.14
geopy==2.2.0
boto3==1.24.66